Rewrite
Rewrote my code for numberofevents.test.js and numberofevents.js because after trying to do integration testing realized that the testing i wrote was not adequate and causing issues. Main issue was it was assigning the default value of 32 to the display rather than the default number in the box being 32. So when trying to replace the value in the box it was throwing errors. also had to update labeling in numberofevents.js to not cause issues with other components that were written.